Extract the CSS AST class-name fetching to a new class
